baseline overfitting exp.

+++ diffuser_baselines/models/diffuser_navigator.py
@@ -206,7 +206,7 @@ class DiffusionNavigator(nn.Module):
 
         # print(f"shapes : {pred.shape} ; {noising_timesteps.shape} ; {noised_orc_action_tokens.shape}")
         step_out = self.noise_scheduler.step(
-            torch.randn(pred[0].unsqueeze(0).shape, device=oracle_action_tokens.device), torch.randn(noising_timesteps[0].shape, device=oracle_action_tokens.device), noised_orc_action_tokens[0].unsqueeze(0)
+            torch.randn(pred[0].unsqueeze(0).shape, device=oracle_action_tokens.device), torch.randn(noising_timesteps[0].shape, device=oracle_action_tokens.device).long(), noised_orc_action_tokens[0].unsqueeze(0)
         )
 
         denoised = step_out["prev_sample"]
